New function in chemical_potential_free_energy.py
Added a function in chemical_potential_free_energy.py to compute the average concentration and average chemical potential of the system. Before this operations were performed directly in the simulation.py file, now they are executed through a call of the new function.

Added also a test to verify that the new function works properly.